% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buc_baneasa}
\alias{buc_baneasa}
\title{Bucharest Baneasa daily climate data}
\format{A data frame with 365 rows and 9 variables:
\itemize{
  \item CODGE:    geographical ID
  \item Date:     date field
  \item Precip:   sums of precipitation  (mm)
  \item Tavg:     mean temperature (°C)
  \item Tmax:     maximum temperature (°C)
  \item Ws:       mean wind speed (°m/s)
  \item Rh:       relative humidity (\%)
}}
\source{
METEO ROMANIA \url{http://www.meteoromania.ro/}
}
\usage{
data(buc_baneasa)
}
\description{
Daily climate data measured in 2015 at Bucuresti Baneasa weather station.
}
\examples{
data(buc_baneasa)
str(buc_baneasa)
hist(buc_baneasa$Precip)



}
\keyword{datasets}

