% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_0.R
\name{et_0}
\alias{et_0}
\title{The FAO Penman-Monteith crop reference evapotranspiration}
\usage{
et_0(Lon, Lat, Dates, Tavg, Rh, Sd, Ws)
}
\description{
The panel of experts recommended  in May 1990 the adoption of the
FAO Penman-Monteith  combination method as a new standard for reference
evapotranspiration and  advised on procedures for calculation of the various
 parameters (http://www.fao.org/docrep/X0490E/x0490e06.htm)
 @param Lon longitude in degrees (WGS84; EPSG:4326)
 @param Lat latitude in degrees (WGS84; EPSG:4326)
 @param Dates date field reprezenting days of measurements
 @param Tavg daily mean temperature (°C)
 @param Rh daily mean relative humidity (%)
 @param Sd daily sunshine duration (hours)
 @param Ws daily wind speed (m/s)
 @references http://www.fao.org/docrep/X0490E/x0490e08.htm
}

