% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gpm.R
\name{get_gpm}
\alias{get_gpm}
\title{Download GPM (IMERG) H5 files to local directory}
\usage{
get_gpm(product, begin = NULL, end = NULL)
}
\arguments{
\item{product}{Character. Contents of the IMERG output.}

\item{begin}{Character. Subset begin date.}

\item{end}{Character. Subset end date.}
}
\description{
Download GPM (IMERG) H5 files to local directory
}

