% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpm.R
\name{read_gpm}
\alias{read_gpm}
\title{Read GMP data to raster file}
\usage{
read_gpm(file, dataset, extent = NA)
}
\arguments{
\item{file}{cclea catre fisierul h5}

\item{dataset}{tipul de dataset care se doreste a se citi}

\item{extent}{of the gridfile to be read in R, a vector of a length 4; order= xmin, xmax, ymin, ymax)}
}
\value{
raster object with layer names from dataset
}
\description{
Read GMP data to raster file
}
\details{
Fisierele sunt desarcate de aici ftp://jsimpson.pps.eosdis.nasa.gov/data/imerg/early/
}
\examples{
# Read h5 gmp file to raster
require(raster)
r <- read_gpm(file=system.file("extdata/3B-HHR-E.MS.MRG.3IMERG.20151001-S000000-E002959.0000.V03E.RT-H5",
package = "climatetools"),dataset="precipitationCal")
plot(r)
}

