% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGPM.R
\name{readGPM}
\alias{readGPM}
\title{Read GMP data to raster file}
\usage{
readGPM(file, dataset)
}
\arguments{
\item{file}{cclea catre fisierul h5}

\item{dataset}{tipul de dataset care se doreste a se citi}
}
\value{
raster object with layer names from dataset
}
\description{
Read GMP data to raster file
}
\details{
Fisierele sunt desarcate de aici ftp://jsimpson.pps.eosdis.nasa.gov/data/imerg/early/
}
\examples{
# Read h5 gmp file to raster
r<-readGPM(file=system.file("extdata/3B-HHR-E.MS.MRG.3IMERG.20151001-S000000-E002959.0000.V03E.RT-H5",
package = "climatetools"),dataset="precipitationCal")
plot(r)
}

