% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_log.R
\name{wind_log}
\alias{wind_log}
\title{Extrapolate the wind speed at a certain height by using the log law}
\usage{
wind_log(Vref, Zref, Z0, Z)
}
\arguments{
\item{Vref}{known velocity at height Zref}

\item{Zref}{reference height where vref is known}

\item{Z0}{roughness length in the current wind direction}

\item{Z}{height above ground level for velocity to be calculated}
}
\description{
Extrapolate the wind speed at a certain height by using the log law
}

