% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_stats.R
\name{wind_stats}
\alias{wind_stats}
\title{Weibull parameters and wind speed characteristics}
\usage{
wind_stats(x, rho = 1.225)
}
\arguments{
\item{x}{a numeric vector containing hourly wind speed data}

\item{rho}{air density. Default value is 1.225 kgm^−3 for a
temperature of 15°C and a standard sea level pressure of 1013.25 hPa}
}
\value{
A list with four components:
\itemize{
\item mean value:
\item max. power wind speed:
\item wind power density:
\item A scale parameter
\item k shape parameter
}
}
\description{
Weibull parameters and wind speed characteristics
}
\details{
Calculate some simple characteristics
of a Weibull distribution function, given the scale parameter A and the
shape parameter k. The Weibull characteristics calculated are (Troen and Petersen, 1989):
\itemize{
\item mean value:
\item max. power wind speed:
\item wind power density:
}
}
