% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpm.R
\name{read_gpm}
\alias{read_gpm}
\title{Read GMP data to raster file}
\usage{
read_gpm(file, dataset = "precipitationCal", extent = NULL)
}
\arguments{
\item{file}{a character string naming a H5 file}

\item{dataset}{to be read from H5 file; see details}

\item{extent}{of the gridfile to be read in R, a vector of a length 4;
order= xmin (-180), xmax (180), ymin (-90), ymax (90)}
}
\value{
raster object with layer names from dataset
}
\description{
Read GMP data to raster file
}
\details{
In order to download data from the PPS FTPs you must first register your email
address with the Precipitation Processing System, using this page:
 http://registration.pps.eosdis.nasa.gov/registration/.

Fisierele sunt desarcate de aici ftp://jsimpson.pps.eosdis.nasa.gov/data/imerg/early/

Data sets:

\itemize{
\item \code{HQobservationTime}: precipitation-relevant satellite passive microwave
source time (min. into half hour);
\item \code{HQprecipitation}: precipitation-relevant satellite passive microwave
source (mm/hr);
\item \code{HQprecipSource}: precipitation-relevant satellite passive microwave
source sensor identifier (index values);
\item \code{IRkalmanFilterWeight}: Kalman filter weight for IR (percent);
\item \code{IRprecipitation}: (mm/hr);
\item \code{precipitationCal}: snapshot precipitation – calibrated (mm/hr);
\item \code{probabilityLiquidPrecipitation}: probability of liquid precipitation phase.
}
}
\examples{
# Read h5 gmp file to raster
require(raster)
r <- read_gpm(file=system.file("extdata/3B-HHR-E.MS.MRG.3IMERG.20151001-S000000-E002959.0000.V03E.RT-H5",
package = "climatetools"),dataset="precipitationCal", extent=c(20,30,43,49))
plot(r)
}

