% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_log.R
\name{wind_log}
\alias{wind_log}
\title{Extrapolate the wind speed at a certain height by using the log law}
\usage{
wind_log(Vref, Zref, Z0, Z)
}
\arguments{
\item{Vref}{known velocity at height Zref}

\item{Zref}{reference height where Vref is known}

\item{Z0}{roughness length in the current wind direction}

\item{Z}{height above ground level for velocity to be calculated}
}
\value{
Wind speed values computed at certain height (Z)
}
\description{
This function extrapolates the wind speed at a certain height
by using the log law. The increase or decrease of wind speed as a function of
height above ground can be computed by this logarithmic expression.
}
\references{
Tennekes, H. "The logarithmic wind profile." Journal of the Atmospheric Sciences 30, no. 2 (1973): 234-238.
}
